brcmstb_nand: Changing brcmstb_nand_intc to brcmstb_nand_soc
Originally we thought brcmstb_nand_intc is only going to contain
interrupt related customizations, which is why it was called "_intc".
But now we realize there's going to be more than just the interrupt
control logic. For example, on certain iProc SoCs, one needs to toggle
a register to configure the endianness of the APB bus access before
accessing the flash cache data registers. Therefore, it's more
appropriate to call it brcmstb_nand_soc

Signed-off-by: Ray Jui <rjui@broadcom.com>